+++ NaoCar/Modules/AutoDrive/AutoDrive.cpp
@@ -257,7 +257,7 @@ int		AutoDrive::thread()
 	&& lineSpace(collapsedLines[0], redLine) < redLineLength) {
 	Vec4i nearest = collapsedLines[0];
 	double minSpace = lineSpace(collapsedLines[0], redLine);
-	for (int i = 1; i < collapsedLines.size(); ++i) {
+	for (size_t i = 1; i < collapsedLines.size(); ++i) {
 	  double space = lineSpace(collapsedLines[i], redLine);
 	  // If the lines are not too far from each other
 	  if (space < minSpace && (space < redLineLength)) {
@@ -288,7 +288,7 @@ int		AutoDrive::thread()
 	double distanceToPrevX = -1;
 	double newX = -1;
 	for(size_t i = 0; i < collapsedLines.size(); i++ ) {
-	  if ((collapsedLines[i][1] >= y && collapsedLines[i][3] <= y) || (collapsedLines[i][1] <= y && collapsedLines[i][3] >= y))
+	  if (((size_t)collapsedLines[i][1] >= y && (size_t)collapsedLines[i][3] <= y) || ((size_t)collapsedLines[i][1] <= y && (size_t)collapsedLines[i][3] >= y))
 	    {
 	      if (collapsedLines[i][2] != collapsedLines[i][0])
 		{
@@ -342,7 +342,7 @@ int		AutoDrive::thread()
     //	 Point(redLine[2], redLine[3]), Scalar(0, 255, 0), 3, 8);
 
     vector<double> roadCoords;
-    for (int value = 380; value < 440; value += 15)
+    for (size_t value = 380; value < 440; value += 15)
       roadCoords.push_back(value * src.rows / 480);
     // roadCoords.push_back(180 * src.rows / 480);
     // roadCoords.push_back(200 * src.rows / 480);
@@ -355,8 +355,8 @@ int		AutoDrive::thread()
     // roadCoords.push_back(400 * src.rows / 480);
     
     vector<Vec2i> road;
-    for (int i = 0; points.size() != 0 && i < points.size() - 1; ++i) {
-      for (int k = 0; k < roadCoords.size(); ++k)
+    for (size_t i = 0; points.size() != 0 && i < points.size() - 1; ++i) {
+      for (size_t k = 0; k < roadCoords.size(); ++k)
 	{
 	  if (points[i][1] == roadCoords[k] && points[i][0] != -1)
 	    {
@@ -379,14 +379,14 @@ int		AutoDrive::thread()
 
     double sum = 0.0;
     double div = 0.0;
-    for (int i = 0; road.size() != 0 && i < road.size(); ++i) {
+    for (size_t i = 0; road.size() != 0 && i < road.size(); ++i) {
       sum += (road.size() - i) * road[i][0];
       div += (road.size() - i);
     }
     if (div > 0)
       sum /= div;
 
-    for (int i = 10; i > 0; --i)
+    for (size_t i = 10; i > 0; --i)
       values[i] = values[i - 1];
     values[0] = sum / src.cols;
     double total = 0;
